Refactor the chatinputfield and relevant files to support display directive in project view
